app/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
groq
tavily-python
tiktoken
chromadb



